+++ tests/testthat/test-state.R
@@ -15,7 +15,11 @@ test_that("peek_options() returns a named list", {
   expect_identical(peek_options("foo", "bar"), list(foo = "FOO", bar = "BAR"))
 })
 
-test_that("is_interactive() can be overridden", {
+test_that("is_interactive() is FALSE when testthat runs", {
+  expect_false(is_interactive())
+})
+
+test_that("is_interactive() honors rlang_interactive option, above all else", {
   expect_true(with_options(rlang_interactive = TRUE, is_interactive()))
   expect_false(with_options(rlang_interactive = FALSE, is_interactive()))
   expect_error(with_options(rlang_interactive = NA, is_interactive()), "must be a single")